feat(compiler): implement pipes and pipe bindings
This commit implements compilation of pipes and pipe bindings within
templates. This includes both a Pipe node type for creation mode, and two
different PipeBindExpr and PipeBindVExpr IR expressions.

A particular challenge was matching an edge case in the way update bindings
are expressed when built by the TemplateDefinitionBuilder, that inverts the
order of two allocations.